products/omega_tutor/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.5.0
python-dotenv>=1.0.0
